*** empty log message ***

+++ zapping/plugins/deinterlace/DI_MoComp2/DI_MoComp2.c
@@ -1,5 +1,5 @@
 /////////////////////////////////////////////////////////////////////////////
-// $Id: DI_MoComp2.c,v 1.3 2005-02-05 22:18:27 mschimek Exp $
+// $Id: DI_MoComp2.c,v 1.3.2.1 2005-05-20 05:45:14 mschimek Exp $
 /////////////////////////////////////////////////////////////////////////////
 // Copyright (c) 2003 Tom Barry & John Adcock.  All rights reserved.
 /////////////////////////////////////////////////////////////////////////////
@@ -25,6 +25,9 @@
 // CVS Log
 //
 // $Log: not supported by cvs2svn $
+// Revision 1.3  2005/02/05 22:18:27  mschimek
+// Completed l18n.
+//
 // Revision 1.2  2005/01/31 07:03:38  mschimek
 // Don\'t define size_t or we run into conflicts with system headers.
 //
@@ -68,7 +71,7 @@
 #include "MoComp2.h"
 //Z #include "..\help\helpids.h"
 
-long SearchEffort = 1;
+int SearchEffort = 1;
 
 ////////////////////////////////////////////////////////////////////////////
 // Start of Settings related code
@@ -113,7 +116,7 @@ DEINTERLACE_METHOD MoComp2Method =
 };
 
 
-DEINTERLACE_METHOD* DI_MoComp2_GetDeinterlacePluginInfo(long CpuFeatureFlags)
+DEINTERLACE_METHOD* DI_MoComp2_GetDeinterlacePluginInfo(int CpuFeatureFlags)
 {
     if (CpuFeatureFlags & FEATURE_SSE2)
     {
